Refactor visualization and data analysis into new analysis package
