+++ setup.py
@@ -28,7 +28,7 @@ setuptools.setup(
     description="Evaluating and comparing reward models.",
     author="DeepMind Technologies Limited and Adam Gleave",
     author_email="adam@gleave.me",
-    python_requires=">=3.6.0",
+    python_requires=">=3.7.0",
     packages=setuptools.find_packages("src"),
     package_dir={"": "src"},
     package_data={"evaluating_rewards": ["py.typed"]},
@@ -42,8 +42,8 @@ setuptools.setup(
         "License :: OSI Approved :: Apache Software License",
         "Programming Language :: Python",
         "Programming Language :: Python :: 3",
-        "Programming Language :: Python :: 3.6",
         "Programming Language :: Python :: 3.7",
+        "Programming Language :: Python :: 3.8",
         "Programming Language :: Python :: Implementation :: CPython",
         "Programming Language :: Python :: Implementation :: PyPy",
     ],
